fix(node): Do not emit fetch spans when tracing is disabled (#13003)
Ensure that TWP works with fetch, but we do not emit a span. Otherwise,
if a user does not use our tracing functionality but adds their own
fetch instrumentation, they get spans twice.

The benefit we have with fetch is that fetch is instrumented via
diagnostics channel, which means that this can be instrumented multiple
times. So if we instrument it, on the plus side this does not interfere
with a user instrumenting it again themselves if they want.

However, with the previous implementation our integration would always
emit spans. If a user is using our own sampler, this is "OK" because it
will sample it out. But if a user does not use our sampler, the spans we
emit there get sent unintentionally.

With this change, we ensure to ignore requests if tracing is disabled.
In order to make TWP work, we instead manually inject the propagator
headers into the fetch request in this scenario.

Part of https://github.com/getsentry/sentry-javascript/issues/12984

Fixes https://github.com/getsentry/sentry-javascript/issues/12969