test(astro): Switch to explicit vitest imports (#13093)
As per https://vitest.dev/config/#globals

> By default, vitest does not provide global APIs for explicitness

I think we should follow vitest defaults here and explicitly import in
the APIs that we need. This refactors our Astro SDK tests to do so.

I also went ahead and fixed up some TS errors in some tests.

This change also removes `environment: 'jsdom'` from the vite config as
it seems nothing needs this for astro. This should means that our tests
are not polluted with jsdom globals, and that future writers have to
explicitly opt-in to the behaviour.