feat(nestjs): Automatic instrumentation of nestjs middleware (#13065)
Adds middleware instrumentation to the `@sentry/nestjs`. The
implementation lives in `@sentry/node` so that both users using
`@sentry/nestjs` directly as well as users still on `@sentry/node`
benefit. The instrumentation is automatic without requiring any
additional setup. The idea is to hook into the Injectable decorator
(every class middleware is annotated with `@Injectable` and patch the
`use` method if it is implemented.

Caveat: This implementation only works for class middleware, which
implements the `use` method, which seems to be the standard for
implementing middleware in nest. However, nest also provides functional
middleware, for which this implementation does not work.