feat(nestjs): Filter all HttpExceptions (#13120)
So far we are filtering exceptions on the status code and do not report
4xx errors. However, we actually only want to capture unexpected
exceptions but all HttpExceptions are only ever thrown explicitly.
Hence, we do not want to capture them and they should be filtered.

In `@sentry/nest` we can use the nest `HttpException` directly, since we
have `@nest/common` as a dependency. In `@sentry/node` this is not the
case, so we filter based on whether a property called status is defined.