ci: Only run affected unit tests on PRs (#13052)
This PR updates our unit test runner on CI to only run unit tests for
packages that have been changed.

In addition, it also updates the list of packages to run in node/browser
only unit test envs - we've been running some in both. We only actually
need to run the fully-browser-only unit tests in the "Browser Unit
Tests" job, the rest (including e.g. core, ...) runs in the node unit
tests in all node versions.

In a follow up step, maybe we can further streamline this and simply run
only unit tests and ensure that e.g. the browser ones only run in one of
the node versions, not all of them. Or find another way that does not
require us to keep 2 lists of packages in separate places. But for now,
this should be an improvement.

I opened a test PR to show this in action, where only something in the
react package was changed:

* Browser Unit Tests:
https://github.com/getsentry/sentry-javascript/actions/runs/10180894649/job/28160138970
* Node Unit Tests:
https://github.com/getsentry/sentry-javascript/actions/runs/10180894649/job/28160141152