feat(nestjs): Automatic instrumentation of nestjs interceptors before route execution (#13153)
Adds automatic instrumentation of interceptors to `@sentry/nestjs`.
Interceptors in nest have a `@Injectable` decorator and implement a
`intercept` function. So we can simply extend the existing
instrumentation to add a proxy for `intercept`.

Remark: Interceptors allow users to add functionality before and after a
route handler is called. This PR adds tracing to whatever happens before
the route is executed. I am still figuring out how to trace any
instructions after the route was executed. Will do that in a separate
PR.