fix(browser): Avoid showing browser extension error message in non-`window` global scopes (#13156)
Relax our browser extension detection check to
avoid showing the error message and blocking `Sentry.init` if the SDK is
not initialized in a `window` global scope. For instance, this will
allow `Sentry.init` to be executed in (service) workers. We likely don't
need to worry about multiple SDK instance collisions in non-`window`
global scopes.