feat(replay): Add "disableSandbox" flag to `logger`
Add an option to toggle using the console sandbox for `logger`. The reason behind this is that we have an experimental option in replay to aid with debugging replays that do not get sent to Sentry. We want to be able to save normal Sentry breadcrumbs so that if we capture exceptions, those logging statements are included in the issue as breadcrumbs.