feat(nestjs): Filter RPC exceptions (#13227)
`RpcExceptions` are always explicitly thrown in nest. Therefore, they
are expected for users and should not be sent to Sentry.

This PR filters these exceptions. In `@sentry/nestjs` we can simply use
`instanceof RpcExceptions` to achieve this. In `@sentry/node` we do not
have access to this class, so we need to check based on a property.

[ref](https://github.com/getsentry/sentry-javascript/issues/13190)