fix(feedback): Ensure feedback can be lazy loaded in CDN bundles (#13241)
This was brought up in slack - if you use a CDN bundle (or the loader)
without feedback included, and you try to lazy-load the
feedbackIntegration, it fails as of today. The reason is that we check
if `window.Sentry.feedbackIntegration` exists, which it _does_, because
we register a shim integration for compatibility in the loader.

So this PR adds a property on the shim integration which we can check
for during lazy loading. While at it, I also added a missing method to
the feedback integration shim.