feat(replay): Add a replay-specific logger
Removes the old `logInfo` function and replaces it with a new replay-specific
logger. Configuration is done when the replay integration is first initialized
to avoid needing to pass around configuration options. This also means
that we cannot select individual log statements to be added as breadcrumbs with
`traceInternals` options.

This also adds a `logger.exception` that wraps `captureException`.

Note that only the following logging levels are supported:
* info
* log
* warn
* error

With two additions:
* exception
* infoTick (needs a better name) - same as `info` but adds the breadcrumb in the next tick due to some pre-existing race conditions

There are two configuration options:
* enable(/disable)CaptureInternalExceptions
* enable(/disable)TraceInternals