feat(node): Add `useOperationNameForRootSpan` to`graphqlIntegration` (#13248)
This introduces a new option for the `graphqlIntegration`,
`useOperationNameForRootSpan`, which is by default `true` but can be
disabled in integration settings like this: `Sentry.graphqlIntegration({
useOperationNameForRootSpan: true })`.

With this setting enabled, the graphql instrumentation will update the
`http.server` root span it is in (if there is one) will be appended to
the span name. So instead of having all root spans be `POST /graphql`,
the names will now be e.g. `POST /graphql (query MyQuery)`.

If there are multiple operations in a single http request, they will be
appended like `POST /graphql (query Query1, query Query2)`, up to a
limit of 5, at which point they will be appended as `+2` or similar.

Closes https://github.com/getsentry/sentry-javascript/issues/13238