feat(sveltekit): Add `wrapServerRouteWithSentry` wrapper (#13247)
Add a wrapper for SvelteKit server routes.
The reason is that some errors (e.g. sveltekit `error()` calls) are not
caught within server (API) routes, as reported in #13224 because in
contrast to `load` function we don't directly try/catch the function
invokation.

For now, users will have to add this wrapper manually. At a later time
we can think about auto instrumentation, similarly to `load` functions
but for now this will remain manual.