fix(astro): Correctly extract request data (#13315)
builds on top of #13306, found while working on #13116

This PR ensures that we correctly extract the request data in our Astro
middleware. Previously we didn't convert the `request.headers` object
into a `Record<string, string>` but simply passed a `Headers` instance.
This caused problems with the `requestDataIntegration` which doesn't
handle the instance correctly.