feat(browser): Add spotlightBrowser integration (#13263)
Adds a browser-side integration for sending events and Sentry requests to Spotlight. 
The integration is not enabled by default but can be added by users if they want 
to explicitly send browser SDK events to spotlight. This is especially helpful if people
use spotlight in the electron app or a standalone browser window instead of the overlay.

---

Co-authored-by: Burak Yigit Kaya <bkaya21@bloomberg.net>