ci: Streamline browser & node unit tests (#13307)
Instead of having to keep to separate lists of include/excludes, we now
keep this in a single list and invert it when necessary.

This way, we should no longer have problems where tests are either run
multiple times, or not in the correct env - just add the test to the
`browser` list in `ci-unit-tests.ts` to make sure it is not run in
multiple node versions unnecessarily. I also added this step to the new
package checklist.