ci: Stabilize CI dependency cache key (#13401)
Ensure it only changes if actual dependencies change.

Previously, we would invalidate the dependency cache every time a
package.json of the workspace changed in any way. This is defensive, but
it also means that we also invalidate if one of these things happen:

1. A script or similar is added/edited for workspace package
2. A release is made, bumping internal dependency versions

This change updates this to calculate the hash with a slightly more
sophisticated approach, which should hopefully ensure we only actually
bust the cache when a dependency _actually_ changes. This should lead to
the dependency cache being re-used much more, because only rarely is an
actual dependency changed.

---------

Co-authored-by: Charly Gomez <charly.gomez@sentry.io>