ci: Update clear cache action to be smarter (#13405)
Previously, we had a CI action to manually clear all caches.

This PR adjusts this so this action can be used in a more granular way:

* By default, the action will now delete caches of any PR runs that are
successful, as well as any caches of release branches.
* You can configure to also delete caches on the develop branch, and/or
to also delete non-successful PR branches.

Additionally, this action will run every midnight, to automatically
clear completed/outdated stuff.

The goal is to keep develop caches as long as possible, and clear out
other caches, unless they failed (which indicates you may want to re-run
some of the tests) and unless they are currently running (to not break
ongoing tests). Ideally, we do not need to manually run this, but can
rely on automated cleanup over night.