ci: Store build output as artifacts instead of cache (#13487)
This PR changes it so that the build output is kept as an artifact, not
a cache. This way, this should never be lost on us.

We keep the NX cache as before.

I chose a retention period of 7 days, which means that after 7 days you
could no longer re-run a workflow partially. IMHO that's a reasonable
start, we can adjust this if needed.