ci: Do not fail on missing dependency cache (#13492)
This PR updates CI to never fail on cache misses. Instead, we rerun
install or rebuild the tarballs, if necessary.

One tricky aspect of this is that `yarn install` will fail when running
on node 14, because some dependencies do not work with it. We "fix" this
by temporarily installing the default node version in this case, run
`yarn install`, then revert to node 14.