core(lint): Add custom `no-focused-tests` and `no-skipped-tests` rules (#13461)
Adds two simple custom rules to ignore
`(it|test|describe).(skip|only)`. These rules now also flag vitest-based
`skip` and `only` functions but led to duplications with the two rules
from `eslint-plugin-jest`. So this patch also disables the jest versions in
favour of the custom rules. To be clear, the custom rules are likely a
bit less robust than the jest/vitest version but until we can use the
actual vitest plugin, I think it's fine to stay with our custom version.