fix(nestjs): Check arguments before instrumenting with `@Injectable` (#13544)
Before this fix, calling a `use` method on a service that does not
implement the `NestMiddleware` interface resulted in an error. This is
because we try to proxy the third argument to the function, which in
middleware is an express `NextFunction`, but in other classes can be
anything, potentially even undefined.

This fix introduces further guards to early-return in non-middleware targets.

Added a test to verify that services with `use` work fine now. Also
added additional tests to verify that this behavior does not occur for
`canActivate` (guard), `intercept` (interceptor) and `transform` (pipe)
methods.