fix(utils): Keep logger on carrier (#13570)
Today, if you have a pluggable integration, it will incorrectly use it's
own logger, which will generally be disabled. This is because we inline
all the logger code into the bundle, but the logger is never enabled.

Now, the logger is kept as a singleton on the carrier, ensuring that we
always use the same logger.

I added browser integration tests to cover this - especially one in
feedback which failed before.

NOTE: This is still not really ideal, because we still include the whole
`makeLogger` code in each pluggable integration that uses the logger 🤔
but we can look into this later - now, at least logging should work for
pluggable integrations.