feat(nuxt): Add server config to root folder (#13583)
> This is a draft PR as this approach leads to an error as `hook.mjs` is
not included in the `node_modules`. This has been fixed upstream but was
not yet released for nuxt.

Makes it possible to include a `sentry.server.config.ts` file in the
root folder alongside `sentry.client.config.ts`. Currently, it has to be
added in the `public` folder which is not 100% ideal.