feat(otel): Upgrade @opentelemetry/semantic-conventions to 1.26.0
In 1.26.0 otel-js has updated the deprecations for the attributes based
on the new changes to semantic conventions.

They also changed the name of some exports, for example:
`SEMATTRS_HTTP_ROUTE` -> `ATTR_HTTP_ROUTE`. Some exports names were not
able to be changed because they are imported from a subpath export
@opentelemetry/semantic-conventions/incubating. This subpath breaks some
bundling setups, so we are unable to use it.