ref(profiling): Conditionally shim cjs globals (#13267)
The shims should only be applied if the globals are not present, else it
results in double decl and a runtime error. The profiling SDK should
gracefully handle env where the shims are already provided.

I couldn't find a way to modify the shim as it is hardcoded in the
plugin we are using so I went with the replace plugin approach and a
placeholder value #poormansmacros.