fix(node): Remove ambiguity and race conditions when matching local variables to exceptions (#13501)
Closes #13415

This PR only modifies the async version of this integration which is
used for Node > v19. I tried applying similar changes to the sync
integration and I cannot get it to work without causing memory leaks.

@Bruno-DaSilva has been helping me explore different ways to fix a few
fundamental issues with the local variables integration. Bruno found a
way to [write to the error
object](https://github.com/getsentry/sentry-javascript/issues/13415#issuecomment-2313754556)
from the debugger which removes any ambiguity over which variables go
with which exception.

This allows us to remove the stack parsing and hashing which we were
using previously to match up exceptions.

Rather than write the `objectId` to the error, I have used this to write
the entire local variables array directly to the error object. This
completely negates the need to post the local variables from the worker
thread which removes any possibility of race conditions. We then later
pull the local variables directly from
`hint.originalException.__SENTRY_ERROR_LOCAL_VARIABLES__`.