fix(browser): Ensure Standalone CLS span timestamps are correct (#13649)
Fix a bug in the initial experimental CLS standalone span
implementation. Previously we'd add the CLS start timestamp value in ms
to the performance time origin timestamp which was already converted to
seconds. Ensure that we first add time origin and the CLS start timestamp 
and then convert to seconds

---------

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>