feat(replay): Add experimental option to allow for a checkout every 6 minutes (#13069)
Including more checkouts will improve replayer scrubbing since it will
reduce the number of mutations that need to be processed (especially for
longer replays). The downside is that it will increase the size of
replays since we will have up to 9 more snapshots per replay (max replay
duration is 60 minutes / 6 minute checkouts).