fix(node): Don't overwrite local variables for re-thrown errors (#13644)
Another way to fix this issue would be to check via the debugger if the
property exists already on the error object and bail early before we
have local variables. However, this would add an extra round trip call
via the debugger for every error. Since re-thrown errors are far less
likely, I decided instead to just not overwrite existing local
variables.

This PR also adds a `Debugger.resume` call in the catch case to ensure
that they debugger will always resume even if we get errors while
debugging.

It's worth noting that this only fixes the issue in Node v20+ where we
use the async debugging interface.