ref(browser): Move navigation span descriptions into op (#13527)
Makes some modifications to `browser` spans:

- Moves the description of navigation related `browser` spans into the
op, e.g. `browser - cache` -> `browser.cache`
- Sets the description to the `performanceEntry` objects' `names`, in
this context it is the URL of the page

This change is being made so that these `browser` spans can be ingested
and grouped. Currently, all browser spans are grouped into a singular
`browser` span, despite each of the operations that these span represent
doing very different things.

This will improve the experience in the Spans tab of transaction summary
and span summary, since we will be able to have proper groupings for
`browser` spans.