fix(node): Ensure node-fetch does not emit spans without tracing (#13765)
Found this while working on
https://github.com/getsentry/sentry-javascript/pull/13763.

Oops, the node-otel-without-tracing E2E test was not running on CI, we
forgot to add it there - and it was actually failing since we switched
to the new undici instrumentation :O

This PR ensures to add it to CI, and also fixes it. The main change for
this is to ensure we do not emit any spans when tracing is disabled,
while still ensuring that trace propagation works as expected for this
case.

I also pulled some general changes into this, which ensure that we patch
both `http.get` and `http.request` properly.