fix(core): Remove `sampled` flag from dynamic sampling context in Tracing without Performance mode (#13753)
Fix a bug in Core that surfaced in Node apps configured for
Tracing without Performance.

Previously, we'd incorrectly add `sampled: false` flag when creating the
DSC from an active span if the application was configured for TwP. This
is possible because in TwP, Otel still emits non-recording spans for the
incoming requests. Our implementation in Core didn't account for this
edge case yet.