feat(replay): Do not log "timeout while trying to read resp body" as exception (#13965)
Change this to be a warn level log instead of an exception (which gets
captured to Sentry). Since this is an error we are throwing ourselves
and it is to be expected, no need to treat as an exception.

This also adds a new meta string to differentiate from an error while
parsing. This means we can show a more specific error message on the
frontend.