test(nuxt): Use `sentry.server.config.ts` in E2E tests (#13999)
The E2E tests still used an `instrument` file in the `public` folder
(which is not the way to setup the SDK anymore) because
`import-in-the-middle/hook.mjs` was not available and threw an error. To
be able to properly test the setup with `sentry.server.config.ts`, a
function was added that copies the `import-in-the-middle` folder to the
build output to include all files.