feat(nuxt): Respect user-provided source map generation settings (#14020)
Nuxt implementation for:
https://github.com/getsentry/sentry-javascript/issues/13993
Fixes: https://github.com/getsentry/sentry-javascript/issues/13997

In Nuxt, there are 3 places to set source maps (and all need to be
enabled):
- `sourcemap`
- `nitro.rollupConfig.output.sourcemap`
- `vite.build.sourcemap`

As Nuxt sets `sourcemap.client: false` ([docs
here](https://nuxt.com/docs/api/nuxt-config#sourcemap)), it's not
possible to determine whether this setting was set by the user or the
framework. Users have to set this manually like this:
```js
export default defineNuxtConfig({
  sourcemap: {
    client: true,
  },
})
```

With this PR, all source maps are set to `'hidden'` if they were
undefined before and keep the setting otherwise. This is done in 3
separate functions (one for vite, rollup and nuxt) to better distinguish
between the settings.